requirements.txt
Django==2.2.5
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
